CheckStyle and compilation corrections

+++ integration-tests/src/test/java/oracle/weblogic/kubernetes/ItLBTwoDomainsNginx.java
@@ -269,13 +269,11 @@ class ItLBTwoDomainsNginx {
    */
   @AfterAll
   void tearDownAll() throws ApiException {
-    if (!SKIP_CLEANUP) {
-      if (pvPvcNamePair != null) {
-        // delete pvc
-        deletePersistentVolumeClaim(pvPvcNamePair.get(1), domainNamespace);
-        // delete pv
-        deletePersistentVolume(pvPvcNamePair.get(0));
-      }
+    if (!SKIP_CLEANUP && pvPvcNamePair != null) {
+      // delete pvc
+      deletePersistentVolumeClaim(pvPvcNamePair.get(1), domainNamespace);
+      // delete pv
+      deletePersistentVolume(pvPvcNamePair.get(0));
     }
     if (nginxHelmParams != null && OKE_CLUSTER) {
       assertThat(uninstallNginx(nginxHelmParams.getHelmParams()))
